requirements.txt
numpy
cupy-cuda11x
matplotlib
flopth
